Moved src/types.js to types.js, and default-types to example-types
Fixes: #3
Fixes: #8
