+++ src/example-types/date.js
