+++ src/example-types/mongoId.js
@@ -1,0 +1,8 @@
+let ObjectID = require('mongodb').ObjectID
+
+module.exports = {
+  hasValue: context => context.data.value,
+  filter: context => ({
+    [context.field]: new ObjectID(context.data.value),
+  }),
+}
