+++ src/example-types/number.js
