+++ test/example-types/text.js
@@ -1,5 +1,5 @@
 let { expect } = require('chai')
-let { text } = require('../src/types')
+let { text } = require('../../types')
 
 describe('text', () => {
   it('should check for values', () => {
